Update dnaadjust.cpp
Since type b uncertainties are assumed at 95% CI (as stated in dsg.typeb file) it is not clear to me if they are reduced to 68% CI before they are added to variances. Not sure if it matters since type b is an arbitrary value but if we follow GAM definition of uncertainty we have two options: standard (68%) and expanded (95%) uncertainty. Since variances are at 68% CI maybe we should add type b also at 68%.